More useful comparison of model times in TestTides
The TimeZone produced by "%Z" with parseTimeOrError is typically fake,
because we're using defaultTimeLocale, so converting model times to
UTC is impossible.

Instead, check that the zone names match and that the LocalTimes are
close. The easiest way to do the latter is to pretend that both
ZonedTimes are in UTC and then compare their UTCTimes.